pyproject.toml
[tool.poetry]
name = "inventree_digikey_integration"
version = "1.0.0"
description = "A CLI to import Digikey parts into an Inventree instance"
authors = ["Eric Udlis <udlis.eric@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
inventree = "^0.13.1"
digikey-api = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
import_digikey_parts = "inventree_digikey_integration.__main__:main"
